NER_for_vc_MVC/requirements.txt
Flask==3.0.2
underthesea
google-api-python-client
google-cloud-bigquery
google-cloud-storage
google-generativeai
keras==2.15.0
tensorflow==2.15.0
numpy
pandas
scikit-learn
datasets
tabulate
db-dtypes
rapidfuzz